исправил замечания из прошлой домашки 4
