+++ niffler_e_2_e_tests_python/README.md
@@ -47,7 +47,7 @@ cd ..
 ```
 Так мы попадем в корень проекта, где тесты лежат
 ```bash
-pytest niffler_e_2_e_tests_python/presentation/authorization/main/profile/test_profile.py niffler_e_2_e_tests_python/presentation/authorization/main/tests_main.py niffler_e_2_e_tests_python/presentation/authorization/test_authorization.py niffler_e_2_e_tests_python/presentation/registration/test_registration.py niffler_e_2_e_tests_python/presentation/test_presentation.py -v --alluredir=allure-result --clean-alluredir --allure-no-capture
+python -m pytest presentation/authorization/main/profile/test_profile.py presentation/authorization/main/tests_main.py presentation/authorization/test_authorization.py presentation/registration/test_registration.py presentation/test_presentation.py -v --alluredir=allure-result --clean-alluredir --allure-no-capture
 ```
 и сама команда для тестов(тут у нас также сгенерируется allure отчет)
 ```bash
@@ -58,7 +58,7 @@ allure serve .\allure-result\
 
 ----
 ```bash
-pytest niffler_e_2_e_tests_python/presentation/authorization/main/profile/test_profile.py niffler_e_2_e_tests_python/presentation/authorization/main/tests_main.py niffler_e_2_e_tests_python/presentation/authorization/test_authorization.py niffler_e_2_e_tests_python/presentation/registration/test_registration.py niffler_e_2_e_tests_python/presentation/test_presentation.py -v
+python -m pytest presentation/authorization/main/profile/test_profile.py presentation/authorization/main/tests_main.py presentation/authorization/test_authorization.py presentation/registration/test_registration.py presentation/test_presentation.py -v
 ```
 
 команда, для обычного запуска автотестов(без allure)
