+++ niffler_e_2_e_tests_python/pages/login_page.py
@@ -1,11 +1,10 @@
 from typing import TYPE_CHECKING
 
 import allure
-
-from niffler_e_2_e_tests_python.configs import AUTH_URL
-from niffler_e_2_e_tests_python.playwright_helper import PlaywrightHelper
-from niffler_e_2_e_tests_python.presentation.presentation_page import PresentationPage
-from niffler_e_2_e_tests_python.utils import get_join_url
+from configs import AUTH_URL
+from pages.presentation_page import PresentationPage
+from utils.playwright_helper import PlaywrightHelper
+from utils.utils import get_join_url
 
 if TYPE_CHECKING:
     from playwright.sync_api import Page
