+++ niffler_e_2_e_tests_python/pages/main_page.py
@@ -1,10 +1,9 @@
 from typing import TYPE_CHECKING
 
 import allure
-
-from niffler_e_2_e_tests_python.configs import FRONT_URL
-from niffler_e_2_e_tests_python.playwright_helper import PlaywrightHelper
-from niffler_e_2_e_tests_python.utils import get_join_url
+from configs import FRONT_URL
+from utils.playwright_helper import PlaywrightHelper
+from utils.utils import get_join_url
 
 if TYPE_CHECKING:
     from playwright.sync_api import Page
