+++ niffler_e_2_e_tests_python/pages/register_page.py
@@ -1,10 +1,9 @@
 from typing import TYPE_CHECKING
 
 import allure
-
-from niffler_e_2_e_tests_python.configs import AUTH_URL
-from niffler_e_2_e_tests_python.playwright_helper import PlaywrightHelper
-from niffler_e_2_e_tests_python.utils import get_join_url
+from configs import AUTH_URL
+from utils.playwright_helper import PlaywrightHelper
+from utils.utils import get_join_url
 
 if TYPE_CHECKING:
     from playwright.sync_api import Page
@@ -30,7 +29,7 @@ class RegisterPage(PlaywrightHelper):
         )
 
     @allure.step
-    def register_new_user(self, username: str, password: str) -> None:
+    def register_user(self, username: str, password: str) -> None:
         """Регистрация пользователя."""
         with allure.step('registering a new user'):
             self.goto_your_page()
