+++ niffler_e_2_e_tests_python/presentation/authorization/conftest.py
@@ -2,14 +2,12 @@ import re
 from typing import TYPE_CHECKING
 
 import pytest
-
-from niffler_e_2_e_tests_python.presentation.authorization.login_page import LoginPage
+from pages.login_page import LoginPage
 
 if TYPE_CHECKING:
+    from pages.main_page import MainPage
+    from pages.presentation_page import PresentationPage
     from playwright.sync_api import Page
-
-    from niffler_e_2_e_tests_python.presentation.authorization.main.main_page import MainPage
-    from niffler_e_2_e_tests_python.presentation.presentation_page import PresentationPage
 
 
 @pytest.fixture(scope='session')
