+++ niffler_e_2_e_tests_python/presentation/authorization/main/conftest.py
@@ -1,15 +1,13 @@
 from typing import TYPE_CHECKING
 
 import pytest
-
-from niffler_e_2_e_tests_python.configs import TEST_PASSWORD, TEST_USER
-from niffler_e_2_e_tests_python.presentation.authorization.main.main_page import MainPage
+from configs import TEST_PASSWORD, TEST_USER
+from pages.main_page import MainPage
 
 if TYPE_CHECKING:
+    from pages.login_page import LoginPage
+    from pages.presentation_page import PresentationPage
     from playwright.sync_api import Page
-
-    from niffler_e_2_e_tests_python.presentation.authorization.login_page import LoginPage
-    from niffler_e_2_e_tests_python.presentation.presentation_page import PresentationPage
 
 
 @pytest.fixture(scope='session')
