+++ niffler_e_2_e_tests_python/presentation/authorization/main/profile/conftest.py
@@ -1,23 +1,19 @@
 from typing import TYPE_CHECKING, Optional
 
 import pytest
-
-from niffler_e_2_e_tests_python.client_api import ClientApi
-from niffler_e_2_e_tests_python.configs import TEST_PASSWORD, TEST_USER
-from niffler_e_2_e_tests_python.fixtures.database import db_niffler_spend  # noqa F401
-from niffler_e_2_e_tests_python.presentation.authorization.main.profile.profile_page import (
-    ProfilePage,
-)
+from configs import TEST_PASSWORD, TEST_USER
+from fixtures.conftest import db_niffler_spend  # noqa F401
+from pages.profile_page import ProfilePage
+from utils.client_api import ClientApi
 
 if TYPE_CHECKING:
     from _pytest.fixtures import SubRequest
     from _pytest.mark import Mark
+    from fixtures.database import DB
+    from pages.login_page import LoginPage
+    from pages.main_page import MainPage
+    from pages.presentation_page import PresentationPage
     from playwright.sync_api import Page
-
-    from niffler_e_2_e_tests_python.fixtures.database import DB
-    from niffler_e_2_e_tests_python.presentation.authorization.login_page import LoginPage
-    from niffler_e_2_e_tests_python.presentation.authorization.main.main_page import MainPage
-    from niffler_e_2_e_tests_python.presentation.presentation_page import PresentationPage
 
 
 @pytest.fixture(scope='session')
