+++ niffler_e_2_e_tests_python/presentation/authorization/main/profile/test_profile.py
@@ -2,15 +2,12 @@ from typing import TYPE_CHECKING
 
 import allure
 import pytest
-
-from niffler_e_2_e_tests_python.configs import TEST_PASSWORD, TEST_USER
-from niffler_e_2_e_tests_python.presentation.authorization.main.profile.profile_page import (
-    ProfilePage,
-)
+from configs import TEST_PASSWORD, TEST_USER
+from pages.profile_page import ProfilePage
 
 if TYPE_CHECKING:
 
-    from niffler_e_2_e_tests_python.fixtures.database import DB
+    from fixtures.database import DB
 
 
 @pytest.fixture
